# Cython performance