… build/lib.linux-x86_64-3.9/test_cy.cpython-39-x86_64-linux-gnu.so
… build/temp.linux-x86_64-3.9/test.o
… test.c
… test.cpp
… test_cy.cpython-39-x86_64-linux-gnu.so