Hi, 

I took the liberty to adding cython to your tests.  Another interesting one would be to compare `numba`. Hope you like it!

The problem set from julia I am getting the following timings

# cython with parallel computation
```
[0.02553773 0.0459578  0.08872843 0.1604135  0.27684927 0.44381857
 0.64729214 0.90215611 1.2150569  0.59125519] # in sec
```
 
 # Julia single threaded
 ```
 running loop on 1 threads 
  0.031174 seconds (8 allocations: 15.259 MiB, 13.99% gc time)
  0.110086 seconds (8 allocations: 61.036 MiB, 3.71% gc time)
  0.253521 seconds (8 allocations: 137.330 MiB, 5.70% gc time)
  0.469424 seconds (8 allocations: 244.141 MiB, 9.96% gc time)
  0.698594 seconds (8 allocations: 381.470 MiB, 6.07% gc time)
  0.954789 seconds (8 allocations: 549.317 MiB, 1.30% gc time)
  1.307076 seconds (8 allocations: 747.681 MiB, 1.24% gc time)
  1.700735 seconds (8 allocations: 976.563 MiB, 1.26% gc time)
  2.115866 seconds (8 allocations: 1.207 GiB, 1.29% gc time)
  2.589587 seconds (8 allocations: 1.490 GiB, 1.58% gc time)
```
  
  # Julia multi-threaded
  
  ```
  running loop on 16 threads 
  0.005241 seconds (84 allocations: 15.266 MiB)
  0.025155 seconds (86 allocations: 61.042 MiB, 20.52% gc time)
  0.047858 seconds (87 allocations: 137.336 MiB, 11.26% gc time)
  0.087672 seconds (85 allocations: 244.148 MiB, 18.02% gc time)
  0.100579 seconds (86 allocations: 381.477 MiB, 48.64% gc time)
  0.123952 seconds (86 allocations: 549.324 MiB, 35.04% gc time)
  0.114635 seconds (87 allocations: 747.688 MiB, 10.45% gc time)
  0.149740 seconds (85 allocations: 976.570 MiB, 11.71% gc time)
  0.209041 seconds (86 allocations: 1.207 GiB, 10.42% gc time)
  0.238639 seconds (84 allocations: 1.490 GiB, 13.56% gc time)```