@mdmaas 2021-06-22
Nice job! Yes, testing numba would be a great idea, too.